rename ResponsiveChartContainer to ChartContainer
