+++ packages/x-charts/src/ChartContainer/index.ts
@@ -1,0 +1,1 @@
+export * from './ChartContainer';
