+++ packages/x-charts/src/ChartContainer/useChartContainerDimensions.ts
