+++ packages/x-charts/src/LineChart/useLineChartProps.ts
@@ -9,7 +9,7 @@ import { ChartsOnAxisClickHandlerProps } from '../ChartsOnAxisClickHandler';
 import { ChartsOverlayProps } from '../ChartsOverlay';
 import { ChartsTooltipProps } from '../ChartsTooltip';
 import { DEFAULT_X_AXIS_KEY } from '../constants';
-import { ResponsiveChartContainerProps } from '../ResponsiveChartContainer';
+import { ChartContainerProps } from '../ChartContainer';
 import { AreaPlotProps } from './AreaPlot';
 import type { LineChartProps } from './LineChart';
 import { LineHighlightPlotProps } from './LineHighlightPlot';
@@ -61,7 +61,7 @@ export const useLineChartProps = (props: LineChartProps) => {
   const id = useId();
   const clipPathId = `${id}-clip-path`;
 
-  const chartContainerProps: ResponsiveChartContainerProps = {
+  const chartContainerProps: ChartContainerProps = {
     ...other,
     series: series.map((s) => ({
       disableHighlight: !!disableLineItemHighlight,
