+++ packages/x-charts/src/PieChart/PieChart.tsx
@@ -4,10 +4,7 @@ import PropTypes from 'prop-types';
 import { useRtl } from '@mui/system/RtlProvider';
 import { useThemeProps } from '@mui/material/styles';
 import { MakeOptional } from '@mui/x-internals/types';
-import {
-  ResponsiveChartContainer,
-  ResponsiveChartContainerProps,
-} from '../ResponsiveChartContainer';
+import { ChartContainer, ChartContainerProps } from '../ChartContainer';
 import { PieSeriesType } from '../models/seriesType';
 import {
   ChartsTooltip,
@@ -38,10 +35,7 @@ export interface PieChartSlotProps
     ChartsOverlaySlotProps {}
 
 export interface PieChartProps
-  extends Omit<
-      ResponsiveChartContainerProps,
-      'series' | 'leftAxis' | 'bottomAxis' | 'plugins' | 'zAxis'
-    >,
+  extends Omit<ChartContainerProps, 'series' | 'leftAxis' | 'bottomAxis' | 'plugins' | 'zAxis'>,
     Omit<ChartsOverlayProps, 'slots' | 'slotProps'>,
     Pick<PiePlotProps, 'skipAnimation'> {
   /**
@@ -112,7 +106,7 @@ const PieChart = React.forwardRef(function PieChart(inProps: PieChartProps, ref)
   const margin = { ...(isRtl ? defaultRTLMargin : defaultMargin), ...marginProps };
 
   return (
-    <ResponsiveChartContainer
+    <ChartContainer
       {...other}
       ref={ref}
       series={series.map((s) => ({ type: 'pie', ...s }))}
@@ -137,7 +131,7 @@ const PieChart = React.forwardRef(function PieChart(inProps: PieChartProps, ref)
       />
       {!loading && <ChartsTooltip {...tooltip} slots={slots} slotProps={slotProps} />}
       {children}
-    </ResponsiveChartContainer>
+    </ChartContainer>
   );
 });
