--- packages/x-charts/src/ResponsiveChartContainer/index.ts
@@ -1,1 +1,0 @@
-export * from './ResponsiveChartContainer';
